+++ Modules/Frame/localization.lua
@@ -9,6 +9,8 @@ local locales = {
 		bind_on_use_short = "BoU",
 		bind_on_equip_short = "BoE",
 		bind_on_pickup_short = "BoP",
+		new_look = "(new look)",
+		upgrade = "(upgrade)",
 
 		slot_name_error = "Could not get item name for loot slot %s , it may have already been looted. You can turn this message off in /xloot",
 		slot_icon_error = "Could not get item icon for loot slot %s , it may have already been looted. You can turn this message off in /xloot",